Organize feature-prompts classes
