+++ components/feature/prompts/src/main/java/mozilla/components/feature/prompts/dialog/MultiButtonDialogFragment.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.prompts
+package mozilla.components.feature.prompts.dialog
 
 import android.annotation.SuppressLint
 import android.app.Dialog
@@ -11,19 +11,30 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.widget.CheckBox
 import androidx.appcompat.app.AlertDialog
+import mozilla.components.feature.prompts.R
 
 private const val KEY_MANY_ALERTS = "KEY_MANY_ALERTS"
 private const val KEY_USER_CHECK_BOX = "KEY_USER_CHECK_BOX"
+private const val KEY_POSITIVE_BUTTON_TITLE = "KEY_POSITIVE_BUTTON_TITLE"
+private const val KEY_NEGATIVE_BUTTON_TITLE = "KEY_NEGATIVE_BUTTON_TITLE"
+private const val KEY_NEUTRAL_BUTTON_TITLE = "KEY_NEUTRAL_BUTTON_TITLE"
 
 /**
- * [android.support.v4.app.DialogFragment] implementation to display web Alerts with native dialogs.
+ * [android.support.v4.app.DialogFragment] implementation to display a confirm dialog,
+ *  it can have up to three buttons, they could be positive, negative or neutral.
  */
-internal class AlertDialogFragment : PromptDialogFragment() {
+internal class MultiButtonDialogFragment : PromptDialogFragment() {
 
     /**
      * Tells if a checkbox should be shown for preventing this [sessionId] from showing more dialogs.
      */
     internal val hasShownManyDialogs: Boolean by lazy { safeArguments.getBoolean(KEY_MANY_ALERTS) }
+
+    internal val positiveButtonTitle: String? by lazy { safeArguments.getString(KEY_POSITIVE_BUTTON_TITLE) }
+
+    internal val negativeButtonTitle: String? by lazy { safeArguments.getString(KEY_NEGATIVE_BUTTON_TITLE) }
+
+    internal val neutralButtonTitle: String? by lazy { safeArguments.getString(KEY_NEUTRAL_BUTTON_TITLE) }
 
     /**
      * Stores the user's decision from the checkbox
@@ -40,9 +51,7 @@ internal class AlertDialogFragment : PromptDialogFragment() {
             .setTitle(title)
             .setCancelable(true)
             .setMessage(message)
-            .setPositiveButton(android.R.string.ok) { _, _ ->
-                onPositiveClickAction()
-            }
+            .setupButtons()
         return (if (hasShownManyDialogs) addCheckbox(builder) else builder)
             .create()
     }
@@ -50,14 +59,6 @@ internal class AlertDialogFragment : PromptDialogFragment() {
     override fun onCancel(dialog: DialogInterface) {
         super.onCancel(dialog)
         feature?.onCancel(sessionId)
-    }
-
-    private fun onPositiveClickAction() {
-        if (!userSelectionNoMoreDialogs) {
-            feature?.onCancel(sessionId)
-        } else {
-            feature?.onConfirm(sessionId, userSelectionNoMoreDialogs)
-        }
     }
 
     @SuppressLint("InflateParams")
@@ -73,24 +74,38 @@ internal class AlertDialogFragment : PromptDialogFragment() {
         return builder
     }
 
+    private fun AlertDialog.Builder.setupButtons(): AlertDialog.Builder {
+        if (!positiveButtonTitle.isNullOrBlank()) {
+            setPositiveButton(positiveButtonTitle) { _, _ ->
+                feature?.onConfirm(sessionId, userSelectionNoMoreDialogs to ButtonType.POSITIVE)
+            }
+        }
+        if (!negativeButtonTitle.isNullOrBlank()) {
+            setNegativeButton(negativeButtonTitle) { _, _ ->
+                feature?.onConfirm(sessionId, userSelectionNoMoreDialogs to ButtonType.NEGATIVE)
+            }
+        }
+        if (!neutralButtonTitle.isNullOrBlank()) {
+            setNeutralButton(neutralButtonTitle) { _, _ ->
+                feature?.onConfirm(sessionId, userSelectionNoMoreDialogs to ButtonType.NEUTRAL)
+            }
+        }
+        return this
+    }
+
     companion object {
-        /**
-         * A builder method for creating a [AlertDialogFragment]
-         * @param sessionId to create the dialog.
-         * @param title the title of the dialog.
-         * @param message the message of the dialog.
-         * @param hasShownManyDialogs tells if this [sessionId] has shown many dialogs
-         * in a short period of time, if is true a checkbox will be part of the dialog, for the user
-         * to choose if wants to prevent this [sessionId] continuing showing dialogs.
-         */
+        @Suppress("LongParameterList")
         fun newInstance(
             sessionId: String,
             title: String,
             message: String,
-            hasShownManyDialogs: Boolean
-        ): AlertDialogFragment {
+            hasShownManyDialogs: Boolean,
+            positiveButton: String = "",
+            negativeButton: String = "",
+            neutralButton: String = ""
+        ): MultiButtonDialogFragment {
 
-            val fragment = AlertDialogFragment()
+            val fragment = MultiButtonDialogFragment()
             val arguments = fragment.arguments ?: Bundle()
 
             with(arguments) {
@@ -98,10 +113,18 @@ internal class AlertDialogFragment : PromptDialogFragment() {
                 putString(KEY_TITLE, title)
                 putString(KEY_MESSAGE, message)
                 putBoolean(KEY_MANY_ALERTS, hasShownManyDialogs)
+                putString(KEY_POSITIVE_BUTTON_TITLE, positiveButton)
+                putString(KEY_NEGATIVE_BUTTON_TITLE, negativeButton)
+                putString(KEY_NEUTRAL_BUTTON_TITLE, neutralButton)
             }
-
             fragment.arguments = arguments
             return fragment
         }
+    }
+
+    enum class ButtonType {
+        POSITIVE,
+        NEGATIVE,
+        NEUTRAL
     }
 }
