+++ components/feature/prompts/src/test/java/mozilla/components/feature/prompts/file/MimeTypeTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.prompts
+package mozilla.components.feature.prompts.file
 
 import android.content.Context
 import android.content.Intent.ACTION_GET_CONTENT
